Relocation from metafacture-xml.
